Ratings für User werden jetzt wieder geladen
Admins bekommen die durchschnittliche Bewertung

+++ Converter/MultiParameterConverter.cs
@@ -7,7 +7,7 @@ namespace CommitAndForget.Converter
   {
     public object Convert(object[] values, Type targetType, object parameter, System.Globalization.CultureInfo culture)
     {
-      return new Tuple<ImageModel, int>((ImageModel)values[0], (int)values[1]);
+      return new Tuple<ImageModel, decimal>((ImageModel)values[0], (decimal)values[1]);
     }
 
     public object[] ConvertBack(object value, Type[] targetTypes, object parameter, System.Globalization.CultureInfo culture)
